Improve advanced numeric singleton logic
Suppression changes are rare in practice.

The check `#No gauss elimination if r[1] not in rB[[i]]` is now ignored, as the code is useful to see  problems anyway.

The logic following if `(singleton_num[A$r[[j]]][1] & length(A$r[[j]]) > 1)` has been removed, as it is unnecessary after the above change.
Also, there were problems with this since `AnyProportionalGaussInt()` assumes r-values in ordered order.

`doCheck` is now returned instead of `TRUE` in more cases within `AnyProportionalGaussInt()`, ensuring secondary to be transformed to primary.